sxxp

+++ sip/SIPTOOLS/SipToolsServer/src/sipserver/trf/vp/vo/LatVo.java
@@ -11,38 +11,37 @@ package sipserver.trf.vp.vo;
  * @author salim
  */
 public class LatVo {
-    float peak;
-    float avg;
-    float[] latArr;
+    long peak;
+    long avg;
+    long[] latArr;
     JtrVo jitterObj;
 
-    public LatVo(float peak, float avg) {
+    public LatVo(long peak, long avg) {
         this.peak = peak;
         this.avg = avg;
     }
 
-    
-    public float getPeak() {
+    public long getPeak() {
         return peak;
     }
 
-    public void setPeak(float peak) {
+    public void setPeak(long peak) {
         this.peak = peak;
     }
 
-    public float getAvg() {
+    public long getAvg() {
         return avg;
     }
 
-    public void setAvg(float avg) {
+    public void setAvg(long avg) {
         this.avg = avg;
     }
 
-    public float[] getLatArr() {
+    public long[] getLatArr() {
         return latArr;
     }
 
-    public void setLatArr(float[] latArr) {
+    public void setLatArr(long[] latArr) {
         this.latArr = latArr;
     }
 
@@ -53,6 +52,7 @@ public class LatVo {
     public void setJitterObj(JtrVo jitterObj) {
         this.jitterObj = jitterObj;
     }
+
 
     @Override
     public String toString() {
